feat:testleri bir classta topladim
