--- src/test/java/com/easybusticket/tests/PendingPaymentSearchTest_US25.java
@@ -1,46 +1,0 @@
-package com.easybusticket.tests;
-
-import com.easybusticket.pages.AdminDashboardPage;
-import com.easybusticket.pages.AdminPage;
-import com.easybusticket.pages.PendingPaymentPage;
-import lombok.extern.slf4j.Slf4j;
-import org.testng.annotations.Test;
-
-@Slf4j
-public class PendingPaymentSearchTest_US25 extends BaseTestAdmin {
-    @Test
-    public void getDateSearchNoTicket() {
-        //The admin enters a time period in the past when no tickets were received.
-        //log in as an administrator and locate the admin Dashboard.
-        AdminDashboardPage adminDashboardPage = new AdminPage().adminLogin();
-        log.info("Admin logged in " + env);
-        //log in as an administrator and locate the admin Dashboard.
-        adminDashboardPage.clickToPendingPayment();
-        log.info("clicked Pending Payment Page " + env);
-        //PendingPayment enters
-        PendingPaymentPage pendingPaymentPage = new PendingPaymentPage();
-        log.info("Pending Payment entered " + env);
-        //search
-        pendingPaymentPage.searchingTicketNoTicket();
-        log.info("displayed whether a ticket is available or not.  " + env);
-    }
-
-    @Test
-    public void getDateSearchWithTicket() {
-        //The admin enters a name or PR NUMBER
-        //log in as an administrator and locate the admin Dashboard.
-        AdminDashboardPage adminDashboardPage = new AdminPage().adminLogin();
-        log.info("Admin logged in " + env);
-        //log in as an administrator and locate the admin Dashboard.
-        adminDashboardPage.clickToPendingPayment();
-        log.info("clicked Pending Payment Page " + env);
-        //PendingPayment enters
-        PendingPaymentPage pendingPaymentPage = new PendingPaymentPage();
-        log.info("Pending Payment entered " + env);
-        //search
-        pendingPaymentPage.getPNRNumberSearchBoxWithTicket();
-        log.info("displayed whether a ticket is available or not.  " + env);
-
-
-    }
-}
